Also add `@bot ready` reminder for review submitted
